refactor: show players scores
