+++ src/features/game/scores/PlayerSection.tsx
@@ -1,33 +1,51 @@
-import React from 'react';
+import React, { useMemo } from 'react';
 
+import { OPPONENT, PLAYER } from '../../../lib/utils/constants';
+import { useGame } from '../useGame';
 import { LiveScore } from './LiveScore';
 import { PlayerName } from './PlayerName';
 
 type PlayerSectionProps = {
-  avatar: string;
-  altAvatar: string;
-  playerName: string;
-  scores: number;
-  isOpponent: boolean;
+  isOpponent?: boolean;
 };
 
 export const PlayerSection: React.FC<PlayerSectionProps> = ({
-  avatar,
-  altAvatar,
-  playerName,
-  scores,
-  isOpponent,
+  isOpponent = false,
 }) => {
+  const { state } = useGame();
+  const { history, players } = state;
+  const player = isOpponent ? players.opponent : players.player;
+
+  const { playerScore, opponentScore } = useMemo(() => {
+    const playerWins = history.filter(
+      (round) => round.roundResult === PLAYER,
+    ).length;
+
+    const opponentWins = history.filter(
+      (round) => round.roundResult === OPPONENT,
+    ).length;
+    return {
+      playerScore: playerWins,
+      opponentScore: opponentWins,
+    };
+  }, [history]);
+
+  const score = isOpponent ? opponentScore : playerScore;
+
   return (
     <div
       className={`flex ${isOpponent ? 'flex-row-reverse' : 'flex-row'} gap-4`}
     >
       <div>
-        <img src={avatar} alt={altAvatar} className="w-15 h-15" />
+        <img
+          src={player.avatar.imgPath}
+          alt={player.avatar.alt}
+          className="w-15 h-15"
+        />
       </div>
       <div className="flex flex-col gap-1">
-        <PlayerName playerName={playerName} isOpponent={isOpponent} />
-        <LiveScore scores={scores} isOpponent={isOpponent} />
+        <PlayerName playerName={player.name} isOpponent={isOpponent} />
+        <LiveScore scores={score} isOpponent={isOpponent} />
       </div>
     </div>
   );
